Show snackbar above fab in main activity
There is no built-in mechanism for handling edge-to-edge + snackbar + fab on the bottom of the screen together. This would be easier normally if you were to use bottom navigation which would handle window insets, then the fab would anchor to the top of that, and you could set the anchor or the snackbar to the fab. In this case, we run into an issue where if we kept the old behavior of putting the snackbar underneath the fab, the window insets that the fab now has to listen for with edge-to-edge will be added together with the padding for the snackbar. As a result, you get an ugly giant space underneath the fab when the snackbar appears. Additionally, the material guidelines say that we should not be moving buttons like a fab when showing a snackbar.

To resolve this, I centralized the controls for creating a snackbar within the activity which allows us to hold onto a reference to the snackbar and trigger a translation animation whenever the fab changes.
Another thing to note, this can't be done purely by changing the anchor of the snackbar to the fab and setting the fab to View.GONE whenever it is hidden. The snackbar will still appear in the same place as the fab when it's visible. You also can't change the position of the snackbar by changing the anchor after it has been shown. It will always appear in the position of the original anchor.